requirements.txt
# Dependencias principales
mcp>=0.1.0
fastmcp>=0.7.0
asyncpg>=0.27.0
python-dotenv>=1.0.0

# Desarrollo
typing-extensions>=4.0.0  # Para mejor soporte de tipos
pytest>=7.0.0  # Para pruebas
pytest-asyncio>=0.20.0  # Para pruebas asíncronas

# Calidad de código
black>=23.0.0  # Formateador de código
isort>=5.0.0  # Ordenar importaciones
flake8>=6.0.0  # Linter
mypy>=1.0.0  # Chequeo de tipos estáticos
